Move worker handler code into uploads service.
