+++ internal/codeintel/uploads/worker/mocks_test.go
@@ -22,7 +22,7 @@ import (
 
 // MockDBStore is a mock implementation of the DBStore interface (from the
 // package
-// github.com/sourcegraph/sourcegraph/enterprise/cmd/precise-code-intel-worker/internal/worker)
+// github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/worker)
 // used for unit testing.
 type MockDBStore struct {
 	// DeleteOverlappingDumpsFunc is an instance of a mock function object
@@ -1655,7 +1655,7 @@ func (c DBStoreWithFuncCall) Results() []interface{} {
 
 // MockGitserverClient is a mock implementation of the GitserverClient
 // interface (from the package
-// github.com/sourcegraph/sourcegraph/enterprise/cmd/precise-code-intel-worker/internal/worker)
+// github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/worker)
 // used for unit testing.
 type MockGitserverClient struct {
 	// CommitDateFunc is an instance of a mock function object controlling
@@ -2211,7 +2211,7 @@ func (c GitserverClientResolveRevisionFuncCall) Results() []interface{} {
 
 // MockLSIFStore is a mock implementation of the LSIFStore interface (from
 // the package
-// github.com/sourcegraph/sourcegraph/enterprise/cmd/precise-code-intel-worker/internal/worker)
+// github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/worker)
 // used for unit testing.
 type MockLSIFStore struct {
 	// DoneFunc is an instance of a mock function object controlling the
